feat: add default function memory and timeout
