+++ packages/common/src/helpers/extensions.ts
@@ -1,0 +1,14 @@
+export const extensionToContentType = {
+  '.js': 'application/javascript',
+  '.css': 'text/css',
+  '.html': 'text/html',
+  '.png': 'image/png',
+  '.jpg': 'image/jpeg',
+  '.jpeg': 'image/jpeg',
+  '.gif': 'image/gif',
+  '.svg': 'image/svg+xml',
+  '.woff': 'application/font-woff',
+  '.woff2': 'application/font-woff2',
+  '.ttf': 'application/font-ttf',
+  '.otf': 'application/font-otf',
+};
