+++ packages/common/src/index.ts
@@ -1,23 +1,2 @@
-export function envStringToObject(env: { key: string; value: string }[]): Record<string, string> {
-  return env.reduce((acc, { key, value }) => {
-    return {
-      ...acc,
-      [key]: value,
-    };
-  }, {});
-}
-
-export const extensionToContentType = {
-  '.js': 'application/javascript',
-  '.css': 'text/css',
-  '.html': 'text/html',
-  '.png': 'image/png',
-  '.jpg': 'image/jpeg',
-  '.jpeg': 'image/jpeg',
-  '.gif': 'image/gif',
-  '.svg': 'image/svg+xml',
-  '.woff': 'application/font-woff',
-  '.woff2': 'application/font-woff2',
-  '.ttf': 'application/font-ttf',
-  '.otf': 'application/font-otf',
-};
+export * from './constants';
+export * from './helpers';
