+++ packages/website/lib/constants.ts
@@ -1,6 +1,3 @@
-export const DEFAULT_MEMORY = 128; // 128MB
-export const DEFAULT_TIMEOUT = 50; // 50ms
-
 export const ORGANIZATION_NAME_MIN_LENGTH = 3;
 export const ORGANIZATION_NAME_MAX_LENGTH = 20;
 export const ORGANIZATION_DESCRIPTION_MAX_LENGTH = 200;
