+++ packages/website/lib/trpc/functionsRouter.ts
@@ -3,7 +3,8 @@ import prisma from '@lagon/prisma';
 import { createRouter } from 'pages/api/trpc/[trpc]';
 import { LOG_LEVELS, TIMEFRAMES } from 'lib/types';
 import { getDeploymentCode, removeDeployment } from 'lib/api/deployments';
-import { DEFAULT_MEMORY, DEFAULT_TIMEOUT, FUNCTION_NAME_MAX_LENGTH, FUNCTION_NAME_MIN_LENGTH } from 'lib/constants';
+import { FUNCTION_NAME_MAX_LENGTH, FUNCTION_NAME_MIN_LENGTH } from 'lib/constants';
+import { FUNCTION_DEFAULT_MEMORY, FUNCTION_DEFAULT_TIMEOUT } from '@lagon/common';
 import * as trpc from '@trpc/server';
 
 export const functionsRouter = () =>
@@ -224,8 +225,8 @@ export const functionsRouter = () =>
                 })),
               },
             },
-            memory: DEFAULT_MEMORY,
-            timeout: DEFAULT_TIMEOUT,
+            memory: FUNCTION_DEFAULT_MEMORY,
+            timeout: FUNCTION_DEFAULT_TIMEOUT,
             env: {
               createMany: {
                 data: input.env.map(({ key, value }) => ({
